Move offcanvas function to window object
